# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PC Builder API",
    "contact": {
      "name": "Juozas Petryla"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "Naudotojai"
        ],
        "summary": "Gauti naudotojų sąrašą",
        "description": "Tik administratorius gali skaityti puslapiuojamą paskyrų sąrašą ir blokavimo būsenas.",
        "operationId": "list_users_api_v1_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2147483647,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserAdminRead"
                  },
                  "title": "Response List Users Api V1 Users Get"
                }
              }
            }
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}": {
      "get": {
        "tags": [
          "Naudotojai"
        ],
        "summary": "Gauti viešą naudotojo informaciją",
        "description": "Prisijungusiems grąžina tik naudotojo ID ir viešą vardą.",
        "operationId": "get_public_user_api_v1_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Naudotojai"
        ],
        "summary": "Pašalinti naudotoją",
        "description": "Tik administratorius; pašalina paskyrą ir sesijas, išsaugo anonimizuotą turinį.",
        "operationId": "delete_user_api_v1_users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}/status": {
      "put": {
        "tags": [
          "Naudotojai"
        ],
        "summary": "Blokuoti arba atblokuoti naudotoją",
        "description": "Tik administratorius; blokuojant atšaukiamos visos paskyros sesijos.",
        "operationId": "change_status_api_v1_users__user_id__status_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAdminRead"
                }
              }
            }
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "user",
          "moderator",
          "admin"
        ],
        "title": "Role"
      },
...
      "UserAdminRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "is_blocked": {
            "type": "boolean",
            "title": "Is Blocked"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "role",
          "is_blocked"
        ],
        "title": "UserAdminRead"
      },
      "UserPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username"
        ],
        "title": "UserPublic"
      },
...
      "UserStatusUpdate": {
        "properties": {
          "is_blocked": {
            "type": "boolean",
            "title": "Is Blocked"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "is_blocked"
        ],
        "title": "UserStatusUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}